requirements.txt
Flask
exchangelib
urllib3
python-dotenv
pytest
